adding list of innovations within generation, if the same structire arises within the same generation it gets assigned the same innovation number

+++ neat-dnfs/include/gene.h
@@ -15,9 +15,9 @@ namespace neat_dnfs
 	struct GeneParameters
 	{
 		GeneType type;
-		unsigned long int id;
+		uint16_t id;
 
-		GeneParameters(GeneType type, unsigned long int id)
+		GeneParameters(GeneType type, uint16_t id)
 			: type(type),
 			id(id)
 		{}
